requirements.txt
Django==2.2

django-fsm==2.7.0
django-fsm-admin==1.2.4
djangorestframework==3.11.0
django-rest-swagger==2.2.0
